+++ katarl2/agents/ppo/ppo_cfg.py
@@ -1,17 +1,17 @@
-from typing import Any, Literal, Optional
+from typing import Literal, Optional, Union, Any
 from dataclasses import dataclass
 from katarl2.agents.common.base_agent_cfg import BaseAgentConfig
 
 @dataclass
-class PPOConfig(BaseAgentConfig):
+class PPODiscreteConfig(BaseAgentConfig):
     # Algorithm name
     algo_name: Literal['PPO'] = 'PPO'
     # Policy name
-    policy_name: Literal['Basic'] = 'Basic'
+    policy_name: Literal['Basic', 'Simba'] = 'Basic'
     # Action type
-    action_type: Literal['discrete'] = 'discrete'
+    action_type: Literal['discrete', 'continuous'] = 'discrete'
     # Network name
-    network_name: Literal['CNN+MLP'] = 'CNN+MLP'
+    network_name: Literal['CNN+MLP', 'MLP'] = 'CNN+MLP'
     # Random seed
     seed: int = 42
     # Output train/eval details, level 0,1,2, message from low to high
@@ -68,3 +68,64 @@ class PPOConfig(BaseAgentConfig):
     weight_decay: float = 0.0
     optimizer: Literal['adam', 'adamw'] = 'adam'
 
+
+@dataclass
+class PPOContinuousConfig(PPODiscreteConfig):
+    # Action type
+    action_type: Literal['discrete', 'continuous'] = 'continuous'
+    # Network name
+    network_name: Literal['CNN+MLP', 'MLP'] = 'MLP'
+    # Total environment steps
+    num_env_steps: int = int(1e6)
+    # the learning rate of the optimizer
+    learning_rate: float = 3e-4
+    # the number of steps to run in each environment per policy rollout
+    num_steps: int = 2048
+    # the number of mini-batches
+    num_minibatches: int = 32
+    # the K epochs to update the policy
+    update_epochs: int = 10
+    # the surrogate clipping coefficient
+    clip_coef: float = 0.2
+    # coefficient of the entropy
+    ent_coef: float = 0.0
+
+@dataclass
+class SimbaPPOContinuousConfig(PPOContinuousConfig):
+    # Policy name
+    policy_name: Literal['Basic', 'Simba'] = 'Simba'
+    # the number of critic residual blocks
+    critic_num_blocks: int = 2
+    # the hidden dimension of critic residual block
+    critic_hidden_dim: int = 512
+    # the number of actor residual blocks
+    actor_num_blocks: int = 1
+    # the hidden dimension of actor residual block
+    actor_hidden_dim: int = 128
+
+    # the running mean std for observation
+    obs_rms: Optional[Any] = None
+
+    optimizer: Literal['adam', 'adamw'] = 'adamw'
+    weight_decay: float = 1e-2
+
+@dataclass
+class SimbaPPODiscreteConfig(PPODiscreteConfig):
+    # Policy name
+    policy_name: Literal['Basic', 'Simba'] = 'Simba'
+    # the number of critic residual blocks
+    critic_num_blocks: int = 2
+    # the hidden dimension of critic residual block
+    critic_hidden_dim: int = 512
+    # the number of actor residual blocks
+    actor_num_blocks: int = 1
+    # the hidden dimension of actor residual block
+    actor_hidden_dim: int = 128
+
+    # the running mean std for observation
+    obs_rms: Optional[Any] = None
+
+    optimizer: Literal['adam', 'adamw'] = 'adamw'
+    weight_decay: float = 1e-2
+
+PPOConfig = Union[PPODiscreteConfig, PPOContinuousConfig, SimbaPPOContinuousConfig, SimbaPPODiscreteConfig]