+++ katarl2/envs/wrappers/__init__.py
@@ -1,1 +1,4 @@
 from .repeat_action import RepeatAction
+from .atari_wrappers import apply_atari_wrappers
+from .normalize_observation import NormalizeObservation
+from .normalize_reward import NormalizeReward